Smarter basename system
Basename are no longer hardcoded; now one of two things happen.
* The program is invoked through the kspng script. If that happens,
it reads from the lockfile that the script creates (with the script's
name as its contents) and that is the new basename.
* The program is invoked directly (mono or directly from cmd). In that
case, it finds the assembly location (in System.Reflection), derives
the file name from that, and that is the new basename.